refactor: folder structure and namespace
