+++ MehrakCore.Tests/Services/Commands/Genshin/GenshinImageUpdaterServiceTests.cs
@@ -4,6 +4,7 @@ using System.Text.Json;
 using MehrakCore.ApiResponseTypes.Genshin;
 using MehrakCore.Repositories;
 using MehrakCore.Services.Commands.Genshin;
+using MehrakCore.Services.Commands.Genshin.Character;
 using MehrakCore.Tests.TestHelpers;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
